requirements.txt
python-telegram-bot==20.3
praw==7.7.0
nest-asyncio==1.5.6
requests==2.31.0
aiohttp==3.8.5
nest_asyncio
